+++ src/main/java/com/gbtec/presentation/GenerationPrinter.java
@@ -1,4 +1,4 @@
-package com.gbtec.visualization;
+package com.gbtec.presentation;
 
 import static org.fusesource.jansi.Ansi.ansi;
